+++ src/main/java/com/softeer/podo/admin/mapper/EventMapper.java
@@ -1,13 +1,12 @@
-package com.softeer.podo.admin.model.dto.mapper;
+package com.softeer.podo.admin.mapper;
 
 import com.softeer.podo.admin.model.dto.EventDto;
-import com.softeer.podo.admin.model.dto.EventListResponseDto;
+import com.softeer.podo.admin.model.dto.response.EventListResponseDto;
 import com.softeer.podo.admin.model.dto.EventRewardDto;
 import com.softeer.podo.admin.model.dto.EventWeightDto;
 import com.softeer.podo.admin.model.entity.Event;
 import com.softeer.podo.admin.model.entity.EventReward;
 import com.softeer.podo.admin.model.entity.EventWeight;
-import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
 import java.util.List;
