+++ src/main/java/com/softeer/podo/admin/model/dto/response/EventRewardConfigResponseDto.java
@@ -1,5 +1,7 @@
-package com.softeer.podo.admin.model.dto;
+package com.softeer.podo.admin.model.dto.response;
 
+import com.softeer.podo.admin.model.dto.EventRewardDto;
+import com.softeer.podo.admin.model.dto.EventWeightDto;
 import com.softeer.podo.admin.model.dto.user.UserListDto;
 import jakarta.annotation.Nullable;
 import lombok.AllArgsConstructor;
