+++ src/main/java/com/softeer/podo/event/service/EventLotsService.java
@@ -5,8 +5,13 @@ import com.softeer.podo.admin.model.entity.Role;
 import com.softeer.podo.common.utils.AESUtils;
 import com.softeer.podo.common.utils.URLUtils;
 import com.softeer.podo.event.exception.*;
-import com.softeer.podo.event.model.dto.*;
 import com.softeer.podo.event.model.dto.mapper.LotsEventMapper;
+import com.softeer.podo.event.model.dto.request.LotsApplicationRequestDto;
+import com.softeer.podo.event.model.dto.request.LotsCommentRequestDto;
+import com.softeer.podo.event.model.dto.request.LotsTypeRequestDto;
+import com.softeer.podo.event.model.dto.response.LotsApplicationResponseDto;
+import com.softeer.podo.event.model.dto.response.LotsCommentResponseDto;
+import com.softeer.podo.event.model.dto.response.LotsTypeResponseDto;
 import com.softeer.podo.event.model.entity.LotsComment;
 import com.softeer.podo.event.model.entity.LotsShareLink;
 import com.softeer.podo.event.model.entity.TestResult;
