+++ src/main/java/com/softeer/podo/test/controller/TestController.java
@@ -3,8 +3,8 @@ package com.softeer.podo.test.controller;
 import com.softeer.podo.security.Auth;
 import com.softeer.podo.security.AuthInfo;
 import com.softeer.podo.common.response.CommonResponse;
-import com.softeer.podo.test.model.dto.TestSignUpRequestDto;
-import com.softeer.podo.test.model.dto.TestSignUpResponseDto;
+import com.softeer.podo.test.model.dto.request.TestSignUpRequestDto;
+import com.softeer.podo.test.model.dto.response.TestSignUpResponseDto;
 import com.softeer.podo.test.service.TestService;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
