+++ src/main/java/com/softeer/podo/test/model/dto/request/TestSignUpRequestDto.java
@@ -1,4 +1,4 @@
-package com.softeer.podo.verification.model.dto;
+package com.softeer.podo.test.model.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -11,11 +11,9 @@ import javax.validation.constraints.NotBlank;
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class CheckVerificationRequestDto {
+public class TestSignUpRequestDto {
     @NotBlank
     private String name;
     @NotBlank
     private String phoneNum;
-    @NotBlank
-    private String verificationCode;
 }
