+++ src/main/java/com/softeer/podo/test/service/TestService.java
@@ -3,8 +3,8 @@ package com.softeer.podo.test.service;
 import com.softeer.podo.admin.model.entity.Role;
 import com.softeer.podo.security.jwt.TokenInfo;
 import com.softeer.podo.security.jwt.TokenProvider;
-import com.softeer.podo.test.model.dto.TestSignUpRequestDto;
-import com.softeer.podo.test.model.dto.TestSignUpResponseDto;
+import com.softeer.podo.test.model.dto.request.TestSignUpRequestDto;
+import com.softeer.podo.test.model.dto.response.TestSignUpResponseDto;
 import com.softeer.podo.test.repository.TestUserRepository;
 import com.softeer.podo.admin.model.entity.TestUser;
 import lombok.RequiredArgsConstructor;
