+++ src/test/java/com/softeer/podo/event/service/EventLotsServiceTest.java
@@ -4,7 +4,12 @@ import com.softeer.podo.admin.model.entity.LotsUser;
 import com.softeer.podo.admin.model.entity.Role;
 import com.softeer.podo.admin.repository.LotsUserRepository;
 import com.softeer.podo.event.exception.ExistingUserException;
-import com.softeer.podo.event.model.dto.*;
+import com.softeer.podo.event.model.dto.request.LotsApplicationRequestDto;
+import com.softeer.podo.event.model.dto.request.LotsCommentRequestDto;
+import com.softeer.podo.event.model.dto.request.LotsTypeRequestDto;
+import com.softeer.podo.event.model.dto.response.LotsApplicationResponseDto;
+import com.softeer.podo.event.model.dto.response.LotsCommentResponseDto;
+import com.softeer.podo.event.model.dto.response.LotsTypeResponseDto;
 import com.softeer.podo.event.model.entity.LotsShareLink;
 import com.softeer.podo.event.model.entity.TestResult;
 import com.softeer.podo.event.repository.LotsCommentRepository;
